detect type mismatches in method declarations and definitions

+++ ruamoko/game/GameEntity.r
@@ -15,7 +15,7 @@
 
 @implementation GameEntity
 
-- (BOOL) takeDamage: weapon : inflictor : attacker : damage
+- (BOOL) takeDamage: weapon :inflictor :attacker : (float)damage
 {
 	return NO;
 }
